refactor(settings): move hosted-worker accessors, const, validator to settings_hosted.go (PRD #1021 M3)
